requirements.txt
numpy~=1.18.5
pandas~=1.1.0
setuptools~=46.4.0.post20200518
nibabel~=3.1.1
torch~=1.6.0
Pillow~=7.2.0
tqdm~=4.46.0
monai~=0.2.0
torchvision~=0.7.0
PyYAML~=5.3.1
scikit-learn~=0.23.2
pyellipsoid~=0.0.3
scikit-image~=0.17.2
matplotlib~=3.3.1
tensorboardX~=2.1